fixed the width for the int parameters

+++ ai/ai.h
@@ -6,7 +6,7 @@
 
 /*
 The reason we are using (internally) a 8x8 board is because we are using an
-unsigned long long (so 64 bits long) variable to put the pieces in places. We
+uint64_t (so 64 bits long) variable to put the pieces in places. We
 could use a 7*7 board or even a 7x6 but it would require to rework how the
 pieces are handled, and it may slow down a lot the algorithm.
 */
@@ -27,24 +27,24 @@ pieces are handled, and it may slow down a lot the algorithm.
 class Game {
 private:
     // Multithreading variables
-    int thread_count = 4;
+    int8_t thread_count = 4;
 
     // Players variables
-    int HUMAN = 1;
-    int AI = ~HUMAN; // This way we are sure they are not the same
-    int current_player = HUMAN; // Not used when in mode shared library
+    int8_t HUMAN = 1;
+    int8_t AI = ~HUMAN; // This way we are sure they are not the same
+    int8_t current_player = HUMAN; // Not used when in mode shared library
 
     // The depth of the current move
-    int current_depth = 0;
+    int16_t current_depth = 0;
 
-    int col_ordering[7] = {3, 4, 2, 5, 1, 6, 0};
+    int8_t col_ordering[7] = {3, 4, 2, 5, 1, 6, 0};
 
     // The bitboards for the two players
-    unsigned long long human_board = 0b0;
-    unsigned long long ai_board = 0b0;
+    uint64_t human_board = 0b0;
+    uint64_t ai_board = 0b0;
     uint8_t col_heights[7] = {7, 6, 5, 4, 3, 2, 1};
 
-    int count = 0; // This is used to know how many times something is done
+    int64_t count = 0; // This is used to know how many times something is done
 
     /**
      * @brief Place the new piece on the board
@@ -54,7 +54,7 @@ private:
      * @param heights The list of the heights of the different columns
      * @return Return NOT_ALLOWED if the column is not valid (full or out of bound), or the column otherwise
      */
-    int putPiece(unsigned long long *player, const int col, uint8_t *heights);
+    int8_t putPiece(uint64_t *player, const int8_t col, uint8_t *heights);
 
     /**
      * @brief Remove the piece at the top of the column chosen
@@ -63,7 +63,7 @@ private:
      * @param col The column we want to remove the piece
      * @param heights The list of the heights of the different columns
      */
-    void removePiece(unsigned long long *player, const int col, uint8_t *heights);
+    void removePiece(uint64_t *player, const int8_t col, uint8_t *heights);
 
     /**
      * @brief Count the number of 1 in bitboard
@@ -71,7 +71,7 @@ private:
      * @param bitboard The board representation
      * @return the number of 1
      */
-    int countNbrOne(const unsigned long long bitboard);
+    int8_t countNbrOne(const uint64_t bitboard);
 
     /**
      * @brief Count the number of 4, 3, and 2 aligned disks
@@ -80,7 +80,7 @@ private:
      * @param state The state of the board (if winning or not)
      * @return the score associated with the bitboard
      */
-    int countPoints(const unsigned long long bitboard, bool *state);
+    int8_t countPoints(const uint64_t bitboard, bool *state);
 
     /**
      * @brief Compute the score of the board
@@ -90,7 +90,7 @@ private:
      * @param depth The depth at wich we are looking
      * @return The score computed
      */
-    double evaluateBoard(const unsigned long long bitboard, const unsigned long long oppBitboard, const int depth);
+    double evaluateBoard(const uint64_t bitboard, const uint64_t oppBitboard, const int8_t depth);
 
     /**
      * @brief The negamax variant of the minimax algorithm
@@ -102,7 +102,7 @@ private:
      * @param beta The beta parameter
      * @return The score of the negamax algorithm at this level
      */
-    double negamax(unsigned long long *player, unsigned long long *opponent, uint8_t *heights, double alpha, double beta);
+    double negamax(uint64_t *player, uint64_t *opponent, uint8_t *heights, double alpha, double beta);
 
     /**
      * @brief Apply the minmax algorithm with alpha-beta prunning
@@ -116,7 +116,7 @@ private:
      * @param beta The beta parameter
      * @return The score of the minimax algorithm at this level
      */
-    double minimax(unsigned long long *player, unsigned long long *opponent, uint8_t *heights, const int depth, const bool isMaximising, double alpha, double beta);
+    double minimax(uint64_t *player, uint64_t *opponent, uint8_t *heights, const int8_t depth, const bool isMaximising, double alpha, double beta);
 
     /**
      * @brief Return the best starting move when no evaluation is done
@@ -124,23 +124,23 @@ private:
      * @param heights The list of the heights of the different columns
      * @return the column played
      */
-    int bestStartingMove(const uint8_t *heights);
+    int8_t bestStartingMove(const uint8_t *heights);
 
     /**
      * @brief A structure to copy the important values for multithreading
      * 
      */
     struct value_search {
-        unsigned long long player;
-        unsigned long long opponent;
-        int depth;
+        uint64_t player;
+        uint64_t opponent;
+        int8_t depth;
         uint8_t heights[7];
 
-        value_search(unsigned long long p, unsigned long long o, int d, uint8_t *h) {
+        value_search(uint64_t p, uint64_t o, int8_t d, uint8_t *h) {
             player = p;
             opponent = o;
             depth = d;
-            for (int i = 0; i < 7; i++) {heights[i] = h[i];}
+            for (int8_t i = 0; i < 7; i++) {heights[i] = h[i];}
         }
     };
 
@@ -152,7 +152,7 @@ private:
      * @param best_score A pointer to the variable storing the best score
      * @param best_move A pointer to the variable storing the best move
      */
-    void start_search(value_search values, int column_played, double *best_score, int *best_move);
+    void start_search(value_search values, int8_t column_played, double *best_score, int8_t *best_move);
 
     /**
      * @brief Search the best move the AI can make
@@ -163,7 +163,7 @@ private:
      * @param heights The list of the heights of the different columns
      * @return The number of the column in which we should play
      */
-    int aiSearchMove(unsigned long long *player, unsigned long long *opponent, const int depth, uint8_t *heights);
+    int8_t aiSearchMove(uint64_t *player, uint64_t *opponent, const int8_t depth, uint8_t *heights);
 
 public:
 
@@ -173,7 +173,7 @@ public:
      * @param depth The depth at which the AI should look
      * @return The column in which the AI played or NOT_ALLOWED if the col is not valid
      */
-    int aiMove(const int depth);
+    int8_t aiMove(const int8_t depth);
 
     /**
      * @brief A wrapper around the Human
@@ -181,7 +181,7 @@ public:
      * @param col The column in which the human want to play
      * @return The column in which the Human played or NOT_ALLOWED if the col is not valid
      */
-    int humanMove(const int col);
+    int8_t humanMove(const int8_t col);
 
     /**
      * @brief Reset the state of the board
@@ -200,23 +200,23 @@ public:
      * 
      * @return the status code
      */
-    int run();
+    int8_t run();
 
     /**
      * @brief Get the count object
      * 
      * @return long 
      */
-    int get_count() {
+    int64_t get_count() {
         return count;
     }
 
-    int human_winning();
+    int8_t human_winning();
 
-    int ai_winning();
+    int8_t ai_winning();
 
     bool draw();
 
-    int forceAIMove(const int col);
-    int scoreAIpos();
+    int8_t forceAIMove(const int8_t col);
+    int8_t scoreAIpos();
 };
